- Exhumed: Fixed key offsets. An int can't store a float, effectively truncating to 0 and not holding the 0.5. Thanks MJ :) . Partially addresses #262 - The Torch icon bouncing in the HUD still remains.

+++ source/games/exhumed/src/status.cpp
@@ -641,8 +641,8 @@ private:
         for (int i = nFirstAnim; i >= 0; i = StatusAnim[i].nPrevAnim)
         {
             int nSequence = nStatusSeqOffset + StatusAnim[i].s1;
-            int xoff = StatusAnim[i].s1 == 4 || StatusAnim[i].s1 == 166 ? 0 : 0.5;
-            int yoff = StatusAnim[i].s1 >= 37 && StatusAnim[i].s1 <= 43 ? 0.5 : 0;
+            double xoff = StatusAnim[i].s1 == 4 || StatusAnim[i].s1 == 166 ? 0 : 0.5;
+            double yoff = StatusAnim[i].s1 >= 37 && StatusAnim[i].s1 <= 43 ? 0.5 : 0;
             DrawStatusSequence(nSequence, StatusAnim[i].s2, yoff, xoff);
         }
     }
